requirements.txt
Flask
tensorflow
pillow
werkzeug 